refactor (tests): Tests, cache, response all migrated into submodules.
