+++ src/lib/cache/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "cache"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+priority-queue = { version = "1.2.0" }
+tokio = { version = "1.6.1" }
+async-trait = "0.1.51"
