+++ src/lib/cache/src/lib.rs
@@ -1,0 +1,8 @@
+#[allow(dead_code)]
+mod cache;
+pub use crate::cache::*;
+
+#[cfg(test)]
+mod tests {
+    //TODO migrate tests from cache.rs
+}
