+++ src/lib/response/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "response"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rocket = {version = "0.5.0-rc.1", features=["json"]}
