+++ src/tests/mod.rs
@@ -1,0 +1,5 @@
+pub mod common;
+pub mod rocket;
+pub mod endpoint_account;
+
+pub mod endpoint_conversion;
